implemented marking as soft delete logic
