+++ server/gitrest/packages/gitrest-base/src/routes/index.ts
@@ -44,7 +44,7 @@ export function create(
             blobs: blobs.create(store, fileSystemManagerFactory, repoManagerFactory),
             commits: commits.create(store, fileSystemManagerFactory, repoManagerFactory),
             refs: refs.create(store, fileSystemManagerFactory, repoManagerFactory),
-            repos: repos.create(store, fileSystemManagerFactory, repoManagerFactory),
+            repos: repos.create(store, repoManagerFactory),
             tags: tags.create(store, fileSystemManagerFactory, repoManagerFactory),
             trees: trees.create(store, fileSystemManagerFactory, repoManagerFactory),
         },
