+++ server/gitrest/packages/gitrest-base/src/routes/repository/contents.ts
@@ -6,7 +6,13 @@
 import { handleResponse } from "@fluidframework/server-services-shared";
 import { Router } from "express";
 import nconf from "nconf";
-import { checkSoftDeleted, getRepoManagerParamsFromRequest, IFileSystemManagerFactory, IRepositoryManagerFactory, logAndThrowApiError } from "../../utils";
+import {
+    checkSoftDeleted,
+    getRepoManagerParamsFromRequest,
+    IFileSystemManagerFactory,
+    IRepositoryManagerFactory,
+    logAndThrowApiError,
+} from "../../utils";
 
 export function create(
     store: nconf.Provider,
