+++ server/gitrest/packages/gitrest-base/src/utils/index.ts
@@ -9,4 +9,3 @@ export * from "./helpers";
 export * from "./isomorphicgitManager";
 export * from "./nodeFsManagerFactory";
 export * from "./nodegitManager";
-export * from "./softDeleteUtils";
